tests/wpt/web-platform-tests/web-share/share-securecontext.http.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>WebShare Test: Share from non-secure context</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
        test(() => {
          assert_false('share' in navigator, 'navigator has attribute \'share\'.');
        }, 'navigator.share must be undefined in non-secure context');
    </script>
  </body>
</html>
